Creacion de APIRest

+++ citas-reservacion-sistema-backend/src/main/java/com/femt/citas_reservacion_sistema_backend/service/imp/SedeServiceImp.java
@@ -3,13 +3,15 @@ package com.femt.citas_reservacion_sistema_backend.service.imp;
 import java.util.List;
 import java.util.Optional;
 
-import com.femt.citas_reservacion_sistema_backend.dto.SedeDTO;
+import com.femt.citas_reservacion_sistema_backend.dto.request.SedeRequestDTO;
+import com.femt.citas_reservacion_sistema_backend.dto.response.SedeResponseDTO;
 import com.femt.citas_reservacion_sistema_backend.entity.Sede;
 import org.springframework.stereotype.Service;
 
 import com.femt.citas_reservacion_sistema_backend.mapper.SedeMapper;
 import com.femt.citas_reservacion_sistema_backend.repository.SedeRepository;
 import com.femt.citas_reservacion_sistema_backend.service.SedeService;
+
 
 @Service
 public class SedeServiceImp implements SedeService {
@@ -23,31 +25,36 @@ public class SedeServiceImp implements SedeService {
     }
 
     @Override
-    public List<SedeDTO> listaSedes() throws Exception {
-        return sedeMapper.toDoList(sedeRepository.findAll());
+    public List<SedeResponseDTO> listarSedes() {
+        List<Sede> sedes = sedeRepository.findAll();
+        return sedeMapper.toResponseList(sedes);
     }
 
     @Override
-    public Optional<SedeDTO> obtenerSedePorId(Long idSede) throws Exception {
-        return sedeRepository.findById(idSede).map(sedeMapper::toDTo);
+    public Optional<SedeResponseDTO> obtenerSedePorId(Long id) {
+        return sedeRepository.findById(id)
+                .map(sedeMapper::toResponseDTO);
     }
 
     @Override
-    public void guardarSede(SedeDTO sedeRequest) throws Exception {
-        sedeRepository.save(sedeMapper.toEntity(sedeRequest));
+    public void registrarSede(SedeRequestDTO dto) {
+        Sede sede = sedeMapper.toEntity(dto);
+        sedeRepository.save(sede);
+    }
+
+    @Override
+    public void actualizarSede(Long id, SedeRequestDTO dto) throws Exception {
+        Sede sede = sedeRepository.findById(id)
+                .orElseThrow(() -> new Exception("Sede no encontrada"));
+        sede.setNombre(dto.getNombre());
+        sedeRepository.save(sede);
     }
 
     @Override
     public void eliminarSede(Long id) throws Exception {
+        if (!sedeRepository.existsById(id)) {
+            throw new Exception("Sede no encontrada");
+        }
         sedeRepository.deleteById(id);
-    }
-
-    @Override
-    public Sede actualizarSede(SedeDTO sedeRequest) throws Exception {
-        Sede sede = sedeRepository.findById(sedeRequest.getId())
-                .orElseThrow(() -> new RuntimeException("Sede no encontrada con id: " + sedeRequest.getId()));
-        sede.setId(sedeRequest.getId());
-        sede.setNombre(sedeRequest.getNombre());
-        return sedeRepository.save(sede);
     }
 }
